requirements.txt
cohere==5.13.12
fastapi==0.115.8
langchain==0.3.19
numpy==2.2.3
openai==1.63.2
pinecone==6.0.1
pydantic==2.10.6
PyPDF2==3.0.1
python-dotenv==1.0.1
scikit_learn==1.6.1
uvicorn==0.34.0
